additional moving around
