+++ src/domain/collaboration/authorization/useCollaborationAuthorization.ts
@@ -1,6 +1,6 @@
 import { useCollaborationAuthorizationEntitlementsQuery } from '@/core/apollo/generated/apollo-hooks';
 import { AuthorizationPrivilege, LicenseEntitlementType } from '@/core/apollo/generated/graphql-schema';
-import { useSpace } from '@/domain/space/SpaceContext/useSpace';
+import { useSpace } from '@/domain/space/context/useSpace';
 
 type CollaborationAuthorizationEntitlementsParams = {
   collaborationId: string | undefined;
