+++ src/domain/journey/space/pages/SpaceLayoutSettings/SpaceLayoutSettingsEdit.tsx
@@ -1,6 +1,6 @@
 import PageContentBlock from '@/core/ui/content/PageContentBlock';
 import PageContentColumn from '@/core/ui/content/PageContentColumn';
-import { useSpace } from '../../../../space/SpaceContext/useSpace';
+import { useSpace } from '../../../../space/context/useSpace';
 import InnovationFlowCollaborationToolsBlock from '@/domain/collaboration/InnovationFlow/InnovationFlowDialogs/InnovationFlowCollaborationToolsBlock';
 import useInnovationFlowSettings from '@/domain/collaboration/InnovationFlow/InnovationFlowDialogs/useInnovationFlowSettings';
 import PageContentBlockHeader from '@/core/ui/content/PageContentBlockHeader';
