+++ src/domain/journey/subspace/hooks/useSubSpace.ts
@@ -1,4 +1,4 @@
 import { useContext } from 'react';
-import { SubspaceContext } from '../context/SubspaceProvider';
+import { SubspaceContext } from '../../../space/context/SubspaceProvider';
 
 export const useSubSpace = () => useContext(SubspaceContext);
