+++ src/domain/space/icons/defaultVisuals/Avatar.jpg
