+++ src/domain/space/icons/defaultVisuals/Card.jpg
