+++ src/domain/space/icons/defaultVisuals/defaultVisualUrls.ts
