+++ src/domain/space/layout/tabbedLayout/Tabs/SpaceSubspaces/SubspaceListView.tsx
@@ -15,7 +15,7 @@ import {
   useUpdateTemplateDefaultMutation,
 } from '@/core/apollo/generated/apollo-hooks';
 import { useNotification } from '@/core/ui/notifications/useNotification';
-import { useSpace } from '@/domain/space/SpaceContext/useSpace';
+import { useSpace } from '@/domain/space/context/useSpace';
 import { JourneyCreationDialog } from '@/domain/shared/components/JourneyCreationDialog/JourneyCreationDialog';
 import SubspaceIcon2 from '@/domain/journey/subspace/icon/SubspaceIcon2';
 import { JourneyFormValues } from '@/domain/shared/components/JourneyCreationDialog/JourneyCreationForm';
