+++ src/main/ui/platformSearch/PlatformSearch.tsx
@@ -1,7 +1,7 @@
 import { SelectOption } from '@mui/base';
 import { useTranslation } from 'react-i18next';
 import React, { forwardRef, PropsWithChildren, useCallback, useMemo, useState } from 'react';
-import { useSpace } from '@/domain/space/SpaceContext/useSpace';
+import { useSpace } from '@/domain/space/context/useSpace';
 import SearchBox from '@/core/ui/search/SearchBox';
 import useNavigate from '@/core/routing/useNavigate';
 import { useLocation } from 'react-router-dom';
